* #1: Fixed review findings
