+++ src/test/java/com/exasol/adapter/dialects/hive/HiveSqlDialectTest.java
@@ -28,7 +28,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class HiveSqlDialectTest {
-    private SqlDialect dialect;
+    private HiveSqlDialect dialect;
     private Map<String, String> rawProperties;
 
     @BeforeEach
